minecraft/protocol: Start 1.26.40 (protocol 2168) support
Apply Seb WIP cereal packet updates for Bedrock 1.26.40. Incomplete vs Cloudburst Bedrock_v2168; draft for finishing remaining serializers.

+++ minecraft/protocol/creative.go
@@ -14,7 +14,7 @@ const (
 // icon that represents the group.
 type CreativeGroup struct {
 	// Category is the category the group falls under. It is one of the constants above.
-	Category int32
+	Category byte
 	// Name is the locale name of the group, i.e. "itemGroup.name.planks".
 	Name string
 	// Icon is the item that represents the group in the creative inventory.
@@ -23,7 +23,7 @@ type CreativeGroup struct {
 
 // Marshal encodes/decodes a CreativeGroup.
 func (x *CreativeGroup) Marshal(r IO) {
-	r.Int32(&x.Category)
+	r.Uint8(&x.Category)
 	r.String(&x.Name)
 	r.Item(&x.Icon)
 }
